Fix all /W4 warnings in tests and library headers
C4267 (cluster.h): Add static_cast<int> for size_t->int in centroid
  add/remove calls and the size() return value.

C4101 (test_gaussian/test_gamma): Remove named 'e' from catch clauses
  where the variable was only used inside assert() — a no-op in Release.

C4068/C4996 (test_optimized_matrix3d): Guard GCC diagnostic pragmas
  with #ifndef _MSC_VER; add matching #pragma warning(push/disable:4996)
  for the deprecated-method compatibility test.

C4244 (14 test files): Add static_cast<double> on size_t->Observation
  assignments in ObservationSet loops across training, calculator,
  integration, and linear-ops test files.

C4834 (4 test files): Replace EXPECT_NO_THROW calls that discarded
  [[nodiscard]] return values with substantive assertions (EXPECT_GE,
  EXPECT_EQ, EXPECT_FALSE) that also improve test coverage.

C4189 (15 distribution test files): Add #pragma warning(disable:4189)
  guarded by #ifdef _MSC_VER. These legacy assert()-based test files
  compute variables used only in debug assert() calls; the warning is
  correct but unavoidable without restructuring the tests.

All 36/36 tests pass. Zero warnings in Release build with /W4.

+++ include/libhmm/training/cluster.h
@@ -22,7 +22,7 @@ public:
      * Creates a new Cluster based on one Observation.
      */
     Cluster( Observation o ){
-        centroid.add( o, observations.size( ) );
+        centroid.add( o, static_cast<int>(observations.size()) );
         observations.push_back( o );
     }
 
@@ -30,7 +30,7 @@ public:
      * Adds an observation to the Cluster.
      */
     void onlineAdd( Observation o ){
-        centroid.add( o, observations.size( ) );
+        centroid.add( o, static_cast<int>(observations.size()) );
         observations.push_back( o );
     }
 
@@ -68,7 +68,7 @@ public:
         
         // If observation is found, remove it
         if (i != observations.end()) {
-            centroid.remove( o, observations.size( ) );
+            centroid.remove( o, static_cast<int>(observations.size()) );
             observations.erase( i );
         }
         // Note: If observation not found, we silently ignore it to handle
@@ -93,7 +93,7 @@ public:
      * Returns the number of Observations associated with the Cluster.
      */
     int size( ){
-        return observations.size( );
+        return static_cast<int>(observations.size());
     }
     
     /*
